refactoring some separation of conserns
